Adds configurable default adapter
This patch adds the ability to configure a default adapter on
ActiveModel::Serializer as described in the README:

```ruby
ActiveModel::Serializer.default_adapter = ActiveModel::Serializer::Adapter::HalAdapter
```

The current default adapter, when no other default adapter is specified
(or when the default adapter is set to nil), is set to the NullAdapter
because that's all that we have at the moment to work with. Future work
should change the out-of-the-box default adapter to the `JsonApiAdapter`
when we have this adapter ready.